refactor(admin-games): extract GameSelectorSection to satisfy file-length cap (ARC-710)
CreateGameRoomPage.tsx hit 514 lines after the coming-soon wiring,
breaching the 500-line cap. Moved the game-tile rendering block out
into a sibling GameSelectorSection.tsx component. Drops the page back
under the limit and removes the now-unused TranslationKey import.
